Move the load controller into its own file
